nino: 1

+++ controllers/user.controller.js
@@ -58,6 +58,16 @@ exports.getAllUsers = async (req, res, next) => {
     res.status(500).json({ error: error.message });
   }
 };
+// Lấy tất cả người dùng
+exports.getAllUsersBasic = async (req, res, next) => {
+  try {
+    const sql = `SELECT id, username, name, phone FROM Users`;
+    const [users] = await db.pool.execute(sql);
+    res.json(users);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
 
 // Lấy thông tin người dùng theo ID
 exports.getUserById = async (req, res, next) => {
@@ -81,11 +91,16 @@ exports.getUserById = async (req, res, next) => {
 // Cập nhật thông tin người dùng
 exports.updateUser = async (req, res, next) => {
   const { id } = req.params;
-  const { name, phone, address, gender } = req.body;
+  const { name, phone, address, gender, note } = req.body;
+
+  // Kiểm tra nếu id là 1
+  if (id === '1') {
+    return res.status(403).json({ error: "Thông tin người dùng này là cố định - không thể thay đổi" });
+  }
 
   try {
-    const sql = `UPDATE Users SET name = ?, phone = ?, address = ?, gender = ? WHERE id = ?`;
-    await db.pool.execute(sql, [name, phone, address, gender, id]);  // Thay đổi ở đây
+    const sql = `UPDATE Users SET name = ?, phone = ?, address = ?, gender = ?, note = ? WHERE id = ?`;
+    await db.pool.execute(sql, [name, phone, address, gender, note, id]);
 
     res.json({ message: "User updated successfully" });
   } catch (error) {
@@ -97,15 +112,21 @@ exports.updateUser = async (req, res, next) => {
 exports.deleteUser = async (req, res, next) => {
   const { id } = req.params;
 
+  // Kiểm tra nếu id là 1
+  if (id === '1') {
+    return res.status(403).json({ error: "Không thể xóa người dùng quản trị viên này!" });
+  }
+
   try {
     const sql = `DELETE FROM Users WHERE id = ?`;
-    await db.pool.execute(sql, [id]);  // Thay đổi ở đây
+    await db.pool.execute(sql, [id]);
 
     res.json({ message: "User deleted successfully" });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
 };
+
 
 // Tìm kiếm người dùng theo tên, số điện thoại, địa chỉ, ID
 exports.search = async (req, res, next) => {
